Add a connection-independent pgoutput to ChangeEvent decoder
